fixed test to the correct data type

+++ ERP-CLI.Tests/ProfitPercentTests.cs
@@ -8,7 +8,7 @@ public class ProfitPercentTests
     [InlineData(100, 50, 100)]   // Price double the cost -> 100% profit
     [InlineData(150, 150, 0)]    // Price equal to cost -> 0% profit
     [InlineData(80, 100, -20)]   // Price less than cost -> -20% (loss)
-    public void ProfitPercent_ReturnsExpectedPercentage(double price, double cost, double expected)
+    public void ProfitPercent_ReturnsExpectedPercentage(decimal price, decimal cost, decimal expected)
     {
         Product product = new Product
         {
